fc-companion/backend/requirements.txt
fastapi
uvicorn[standard]
watchdog
sqlalchemy
pydantic
requests
python-dotenv
google-generativeai
